Reverted incident of moving a folder
