+++ app/src/main/res/layout/fragment_home.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Database.Home.HomeFragment">
+    tools:context=".fragments.Home.HomeFragment">
 
     <FrameLayout
         android:layout_width="match_parent"
